[댓글 기능] 닉 미션 제출합니다. (#128)
* [FIX] 로그인하지 않은 User가 게시글을 볼 수 없는 문제 수정

* [ADD] Article, User Entity 간의 다대일 관계 설정 및 관련 test 코드 수정

* [ADD] Article 수정 시 현재 User와 author가 같은지 체크하는 로직 추가

* [ADD] author가 아닌 User가 Article 수정 요청 시 redirect하는 로직 테스트

* [ADD] Comment Entity 생성

* [ADD] Comment 작성 기능 추가

* [ADD] Comment를 Article에 보여주는 기능 구현

* [ADD] 로그인 하지 않은 사용자가 댓글 URL에 접근하는 경우를 interceptor에 추가

* [FIX] 자동 데이터 삽입을 위한 data.sql 설정

* [FIX] article.html의 header 공통부분 수정

* [ADD] 댓글 수정하는 로직 추가

* [ADD] Comment 작성자만 수정, 삭제 버튼이 보이도록 하는 로직 추가

* [ADD] Comment 삭제 기능 구현

* [ADD] Article 삭제 시 해당 Comment를 같이 삭제하는 기능 구현

* [ADD] CommentControllerTests의 Create, Update, Delete Test 생성

* [ADD] CommentServiceTest 추가

* [REF] DTO의 이름을 ~Params에서 ~RequestDto로 변경

* [REF] CommentController의 RequestMapping 중복 제거

* [ADD] Comment 관련 Controller, Service 에 info, debug log 추가

* [FIX] 실패하는 Test 수정 및 저장, 삭제 테스트 통합

* [ADD] error level log 추가

* [REF] Comment의 User fetch전략을 LAZY에서 EAGER로 변경

* [REF] CommentController의 수정, 삭제 uri를 RESTful한 관점으로 수정

* [REF] @Transactional이 필요/불필요한 부분 재정립

* [REF] LoginTestUtil에 검증로직 추가

* [REF] UserControllerTests에서 회원가입, 삭제 test가 setup, teardown에 있으므로 제거

* [FIX] LoginTestUtil의 userSaveRequestDto를 동시사용하는 상황에 발생하는 테스트 문제 수정

* [REF] Article, Comment의 setter제거 및 관련 테스트 수정

* [REF] Article에 isAuthor(user)를 추가하고 관련 로직 수정

* [REF] 가독성을 위해 log 관련 공백라인 추가 및 정리

* [REF] Article, Comment의 update()에 user check 로직 추가

* [REF] domain entity column들에 대하여 nullable=false 설정 추가

* [REF] User의 name, email, password field를 각각 객체로 추출

* [REF] User, UserName, UserEmail, UserPassword에 match 메소드 작성 및 적용

* [REF] ServiceTest에 @Transactional 적용 및 ControllerTest naming 정리

* [REF] Article, Comment Controller Test에 예외관련 Case들 추가

* [ADD] application.yml로 교체 및 .gitignore 수정

* [REF] 각 ControllerTests에 save test 추가

* [REF] Article Delete 로직에서 해당 author인지 체크하는 로직 추가

* [FIX] entity의 삭제 시 참조키 에러가 발생하는 문제를 ondelete 설정으로 해결

* [FIX] entity의 기본 생성자를 public에서 protected로 변경

* [REF] ArticleTest에서 userRepository를 제거하고 ReflectionTestUtils를 사용하여 test

* [REF] ArgumentResolver 사용하여 Session의 user를 미리 추출하도록 변경

* [REF] package 구조 변경: repository -> domain, argumentresolver, interceptor, config -> web

* [REF] Article update 로직을 service에서 entity 내부 update 메소드로 이동

* [REF] Comment update 로직을 service에서 entity 내부 update 메소드로 이동

* [REF] ArticleServiceTest, CommentServiceTest에 author가 아닌 경우 update, delete test 추가

+++ src/main/java/techcourse/myblog/service/ArticleService.java
@@ -4,9 +4,11 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import techcourse.myblog.domain.Article;
-import techcourse.myblog.dto.ArticleSaveParams;
+import techcourse.myblog.domain.User;
+import techcourse.myblog.domain.repository.ArticleRepository;
+import techcourse.myblog.dto.ArticleSaveRequestDto;
 import techcourse.myblog.exception.ArticleNotFoundException;
-import techcourse.myblog.repository.ArticleRepository;
+import techcourse.myblog.exception.IllegalArticleDeleteRequestException;
 
 import javax.transaction.Transactional;
 
@@ -21,26 +23,55 @@ public class ArticleService {
         return articleRepository.findAll();
     }
 
-    public Article save(Article article) {
+    public Article save(ArticleSaveRequestDto articleSaveRequestDto, User author) {
+        Article article = Article.builder()
+                .title(articleSaveRequestDto.getTitle())
+                .coverUrl(articleSaveRequestDto.getCoverUrl())
+                .contents(articleSaveRequestDto.getContents())
+                .author(author)
+                .build();
+
         log.debug("save article={}", article);
         return articleRepository.save(article);
     }
 
     public Article findById(long id) {
         return articleRepository.findById(id)
-                .orElseThrow(() -> new ArticleNotFoundException(ERROR_ARTICLE_NOT_FOUND_MESSAGE));
+                .orElseThrow(() -> {
+                    log.error("error find article by id={}", id);
+                    return new ArticleNotFoundException(ERROR_ARTICLE_NOT_FOUND_MESSAGE);
+                });
     }
 
     @Transactional
-    public void update(ArticleSaveParams articleSaveParams, long id) {
-        log.debug("update article params={}", articleSaveParams);
+    public void update(ArticleSaveRequestDto articleSaveRequestDto, long id, User user) {
+        if (isArticleNotFound(id)) {
+            log.error("update article request by illegal article id={}", id);
+            throw new ArticleNotFoundException(ERROR_ARTICLE_NOT_FOUND_MESSAGE);
+        }
 
         Article article = findById(id);
-        article.update(articleSaveParams);
+        article.update(articleSaveRequestDto, user);
     }
 
-    public void deleteById(long id) {
-        log.debug("delete article id={}", id);
+    @Transactional
+    public void deleteById(long id, User user) {
+        if (isArticleNotFound(id)) {
+            log.error("delete article request by illegal article id={}", id);
+            throw new ArticleNotFoundException(ERROR_ARTICLE_NOT_FOUND_MESSAGE);
+        }
+
+        Article article = findById(id);
+        if (article.isNotAuthor(user)) {
+            log.error("delete article request by illegal user id={}, article id={}", user.getId(), id);
+            throw new IllegalArticleDeleteRequestException();
+        }
+
         articleRepository.deleteById(id);
+        log.debug("delete article id={}", id);
+    }
+
+    private boolean isArticleNotFound(long id) {
+        return findById(id) == null;
     }
 }
